Rearranging
need to make bclim code generic (bmp data -> encoded swizzle)
CLIM is now
