+++ pk3DS.Core/CTR/Images/BFLIM.cs
@@ -1,21 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Runtime.InteropServices;
+using pk3DS.Core.CTR.Images;
 
 namespace pk3DS.Core.CTR
 {
-    public class BFLIM : IXLIM
+    public class BFLIM : BXLIM
     {
-        public byte[] PixelData;
-        public FLIM Footer;
-
-        public uint Magic { get => Footer.Magic; set => Footer.Magic = value; }
-        public ushort Width { get => Footer.Width; set => Footer.Width = value; }
-        public ushort Height { get => Footer.Height; set => Footer.Height = value; }
-        public XLIMEncoding Format { get => Footer.Format; set => Footer.Format = value; }
-
         public BFLIM(Stream data) => ReadBFLIM(data);
         public BFLIM(byte[] data)
         {
@@ -31,336 +21,11 @@ namespace pk3DS.Core.CTR
 
         private void ReadBFLIM(Stream ms)
         {
-            PixelData = new byte[ms.Length - FLIM.SIZE];
+            PixelData = new byte[ms.Length - FLIMHeader.SIZE];
             ms.Read(PixelData, 0, PixelData.Length);
-            var footer = new byte[FLIM.SIZE];
+            var footer = new byte[FLIMHeader.SIZE];
             ms.Read(footer, 0, footer.Length);
-            Footer = footer.ToStructure<FLIM>();
-        }
-
-        /// <summary>
-        /// ARGB 32bpp
-        /// </summary>
-        public byte[] GetImageData(bool crop = true)
-        {
-            var orienter = new BFLIMOrienter(Footer.Width, Footer.Height, Footer.Orientation);
-            uint[] pixels = GetPixels();
-
-            if (!crop)
-            {
-                Footer.Width = (ushort)orienter.Width;
-                Footer.Height = (ushort)orienter.Height;
-            }
-
-            // uint[] -> byte[]
-            byte[] array = new byte[Footer.Width * Footer.Height * 4];
-            for (uint i = 0; i < pixels.Length; i++)
-            {
-                var coord = orienter.Get(i);
-                if (coord.X >= Footer.Width || coord.Y >= Footer.Height)
-                    continue;
-
-                var val = pixels[i];
-                uint o = 4 * (coord.X + coord.Y * Footer.Width);
-                array[o + 0] = (byte)(val & 0xFF);
-                array[o + 1] = (byte)(val >> 8 & 0xFF);
-                array[o + 2] = (byte)(val >> 16 & 0xFF);
-                array[o + 3] = (byte)(val >> 24 & 0xFF);
-            }
-            return array;
-        }
-        public uint[] GetPixels()
-        {
-            return PixelConverter.GetPixels(PixelData, Footer.Format).ToArray();
-        }
-        public byte[] GetPixelsRaw()
-        {
-            var pix = GetPixels();
-            byte[] raw = new byte[pix.Length * 4];
-            Buffer.BlockCopy(pix, 0, raw, 0, raw.Length);
-            return raw;
-        }
-    }
-    [StructLayout(LayoutKind.Sequential, CharSet = CharSet.Ansi)]
-    public struct FLIM : IXLIM
-    {
-        public const int SIZE = 40;
-        public const string Identifier = "FLIM";
-        public bool Valid => LittleEndian && Magic == 0x4D_49_4C_46; // FLIM
-        public bool LittleEndian => BOM == 0xFEFF;
-        public bool BigEndian => BOM == 0xFFFE;
-
-        public uint Magic { get; set; }          // FLIM
-        public ushort BOM;          // 0xFFFE
-        public ushort HeaderLength; // always 0x14
-        public int Version;
-        public uint TotalLength;
-        public uint Count;
-
-        public uint imag; // imag = 67616D69
-        public uint imagLength; // always 0x10
-        public ushort Width { get; set; }
-        public ushort Height { get; set; }
-        public short Alignment;
-        public XLIMEncoding Format;
-        public BFLIMOrientation Orientation;
-        public uint DataSize;
-    }
-
-    public enum XLIMEncoding : byte
-    {
-        L8        = 0x00, // 8    Luminance
-        A8        = 0x01, // 8    Alpha
-        LA4       = 0x02, // 8    Luminance + Alpha
-        LA8       = 0x03, // 16   Luminance + Alpha
-        HILO8     = 0x04, // 16   ?
-        RGB565    = 0x05, // 16   Color
-        RGBX8     = 0x06, // 24   Color
-        RGB5A1    = 0x07, // 16   Color + Alpha
-        RGBA4     = 0x08, // 16   Color + Alpha
-        RGBA8     = 0x09, // 32   Color + Alpha
-        ETC1      = 0x0A, // 4    Color
-        ETC1A4    = 0x0B, // 8    Color + Alpha
-        L4        = 0x0C, // 4    Luminance
-        A4        = 0x0D, // 4    Alpha
-    }
-
-    public static class PixelConverter
-    {
-        private const int BPP_32 = 32;
-        private const int BPP_24 = 24;
-        private const int BPP_16 = 16;
-        private const int BPP_8 = 8;
-        private const int BPP_4 = 4;
-
-        internal static readonly byte[] Convert5To8 = {
-            0x00,0x08,0x10,0x18,0x20,0x29,0x31,0x39,
-            0x41,0x4A,0x52,0x5A,0x62,0x6A,0x73,0x7B,
-            0x83,0x8B,0x94,0x9C,0xA4,0xAC,0xB4,0xBD,
-            0xC5,0xCD,0xD5,0xDE,0xE6,0xEE,0xF6,0xFF
-        };
-
-        public static IEnumerable<uint> GetPixels(byte[] raw, XLIMEncoding e)
-        {
-            int bpp = e.GetBitsPerPixel();
-            if (bpp == BPP_4)
-            {
-                foreach (byte b in raw)
-                {
-                    byte _0 = (byte)(b & 0xF);
-                    byte _1 = (byte)(b >> 4);
-                    yield return GetDecodedPixelValue(_0, e);
-                    yield return GetDecodedPixelValue(_1, e);
-                }
-                yield break;
-            }
-
-            for (int i = 0; i < raw.Length; i += bpp / 8)
-            {
-                uint val = GetEncodedPixelValue(raw, i, bpp);
-                yield return GetDecodedPixelValue(val, e);
-            }
-        }
-
-
-        private static uint GetDecodedPixelValue(uint val, XLIMEncoding e)
-        {
-            byte a = byte.MaxValue, r = 0, g = 0, b = 0;
-            switch (e)
-            {
-                case XLIMEncoding.L4:
-                case XLIMEncoding.L8:
-                {
-                    r = g = b = (byte)val;
-                    break;
-                }
-                case XLIMEncoding.A4:
-                case XLIMEncoding.A8:
-                {
-                    r = g = b = 0xFF;
-                    a = (byte)val;
-                    break;
-                }
-                case XLIMEncoding.HILO8:
-                {
-                    r = (byte)(val >> 8);
-                    g = (byte)(val & 0xFF);
-                    b = byte.MaxValue;
-                    break;
-                }
-                case XLIMEncoding.LA4:
-                {
-                    r = g = b = (byte)(val >> 4);
-                    a = (byte)(val & 0x0F);
-                    break;
-                }
-                case XLIMEncoding.LA8:
-                {
-                    r = g = b = (byte)(val >> 8);
-                    a = (byte)val;
-                    break;
-                }
-                case XLIMEncoding.RGBX8:
-                {
-                    r = (byte)(val >> 16);
-                    g = (byte)(val >> 8);
-                    b = (byte)(val >> 0);
-                    break;
-                }
-                case XLIMEncoding.RGBA8:
-                {
-                    r = (byte)(val >> 24);
-                    g = (byte)(val >> 16);
-                    b = (byte)(val >> 8);
-                    a = (byte) val;
-                    break;
-                }
-                case XLIMEncoding.RGBA4:
-                {
-                    a = (byte)(0x11 * (val & 0xf));
-                    r = (byte)(0x11 * ((val >> 12) & 0xf));
-                    g = (byte)(0x11 * ((val >> 8) & 0xf));
-                    b = (byte)(0x11 * ((val >> 4) & 0xf));
-                    break;
-                }
-                case XLIMEncoding.RGB565:
-                {
-                    r = Convert5To8[(val >> 11) & 0x1F];
-                    g = (byte)(((val >> 5) & 0x3F) * 4);
-                    b = Convert5To8[val & 0x1F];
-                    break;
-                }
-                case XLIMEncoding.RGB5A1:
-                {
-                    r = Convert5To8[(val >> 11) & 0x1F];
-                    g = Convert5To8[(val >> 6) & 0x1F];
-                    b = Convert5To8[(val >> 1) & 0x1F];
-                    a = (val & 1) == 1 ? byte.MaxValue : byte.MinValue;
-                    break;
-                }
-                default:
-                    throw new FormatException($"Unsupported {nameof(XLIMEncoding)} value = {e}");
-            }
-            return (uint)((a << 24) | (r << 16) | (g << 8) | b);
-        }
-
-        public static uint GetEncodedPixelValue(byte[] raw, int offset, int size)
-        {
-            switch (size)
-            {
-                case BPP_32:
-                    return BitConverter.ToUInt32(raw, offset);
-                case BPP_24:
-                    return BitConverter.ToUInt32(raw, offset) & 0x00FFFFFF;
-                case BPP_16:
-                    return BitConverter.ToUInt16(raw, offset);
-                default:
-                    return raw[offset];
-            }
-        }
-        public static int GetBitsPerPixel(this XLIMEncoding e)
-        {
-            if (_32.Contains(e))
-                return BPP_32;
-            if (_24.Contains(e))
-                return BPP_24;
-            if (_16.Contains(e))
-                return BPP_16;
-            if (_8.Contains(e))
-                return BPP_8;
-            return BPP_4;
-        }
-
-        private static readonly HashSet<XLIMEncoding> _32 = new HashSet<XLIMEncoding> 
-        {
-            XLIMEncoding.RGBA8,
-        };
-        private static readonly HashSet<XLIMEncoding> _24 = new HashSet<XLIMEncoding>
-        {
-            XLIMEncoding.RGBX8,
-        };
-        private static readonly HashSet<XLIMEncoding> _16 = new HashSet<XLIMEncoding>
-        {
-            XLIMEncoding.LA8,
-            XLIMEncoding.HILO8,
-            XLIMEncoding.RGB565,
-            XLIMEncoding.RGB5A1,
-            XLIMEncoding.RGBA4,
-        };
-        private static readonly HashSet<XLIMEncoding> _8 = new HashSet<XLIMEncoding>
-        {
-            XLIMEncoding.L8,
-            XLIMEncoding.A8,
-            XLIMEncoding.LA4,
-            XLIMEncoding.ETC1A4,
-        };
-    }
-
-    [Flags]
-    public enum BFLIMOrientation : byte
-    {
-        None = 0,
-        Rotate90 = 4,
-        Transpose = 8,
-    }
-
-    public class BFLIMOrienter
-    {
-        readonly BFLIMOrientation _orientation;
-
-        public uint Width { get; }
-        public uint Height { get; }
-        public uint PanelsPerWidth { get; }
-        
-        public BFLIMOrienter(int width, int height, BFLIMOrientation orientation)
-        {
-            Width = (uint)BCLIM.nlpo2(BCLIM.gcm(width, 8));
-            Height = (uint)BCLIM.nlpo2(BCLIM.gcm(height, 8));
-
-            uint stride = orientation == BFLIMOrientation.None ? Width : Height;
-            PanelsPerWidth = (uint)BCLIM.gcm((int)stride, 8) / 8;
-
-            _orientation = orientation;
-        }
-
-        public Coordinate Get(uint i)
-        {
-            BCLIM.d2xy(i & 0x3F, out uint x, out uint y);
-
-            // Shift Tile Coordinate into Tilemap
-            var tile = i >> 6;
-            x |= (tile % PanelsPerWidth) << 3;
-            y |= (tile / PanelsPerWidth) << 3;
-
-            var coord = new Coordinate(x, y);
-            if (_orientation.HasFlag(BFLIMOrientation.Rotate90))
-                coord.Rotate90(Height);
-            if (_orientation.HasFlag(BFLIMOrientation.Transpose))
-                coord.Transpose();
-            return coord;
-        }
-    }
-
-    public struct Coordinate
-    {
-        public uint X { get; private set; }
-        public uint Y { get; private set; }
-
-        public Coordinate(uint x, uint y)
-        {
-            X = x; Y = y;
-        }
-        public void Transpose()
-        {
-            var tmp = X;
-            X = Y;
-            Y = tmp;
-        }
-        public void Rotate90(uint height)
-        {
-            var tmp = X;
-            X = Y;
-            Y = height - 1 - tmp;
+            Footer = footer.ToStructure<FLIMHeader>();
         }
     }
 }
