+++ pk3DS.Core/CTR/Images/IXLIMHeader.cs
@@ -3,10 +3,14 @@
     /// <summary>
     /// <see cref="BCLIM"/> and <see cref="BFLIM"/> header interface.
     /// </summary>
-    public interface IXLIM
+    public interface IXLIMHeader
     {
         uint Magic { get; set; }
         ushort Width { get; set; }
         ushort Height { get; set; }
+        XLIMEncoding Format { get; set; }
+        XLIMOrientation Orientation { get; set; }
+
+        bool Valid { get; }
     }
 }
