+++ pk3DS.Core/CTR/Images/XLIMOrientation.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace pk3DS.Core.CTR
+{
+    [Flags]
+    public enum XLIMOrientation : byte
+    {
+        None = 0,
+        Rotate90 = 4,
+        Transpose = 8,
+    }
+}
